Add configurable typeforms.
