+++ pages/borrow.tsx
@@ -1,3 +1,4 @@
+import { Survey } from 'features/survey'
 import { serverSideTranslations } from 'next-i18next/serverSideTranslations'
 import React from 'react'
 
@@ -15,6 +16,7 @@ function BorrowPage() {
   return (
     <WithConnection>
       <BorrowView />
+      <Survey for="borrow" />
     </WithConnection>
   )
 }
