+++ pages/earn.tsx
@@ -1,3 +1,4 @@
+import { Survey } from 'features/survey'
 import { serverSideTranslations } from 'next-i18next/serverSideTranslations'
 import { useRouter } from 'next/router'
 import React from 'react'
@@ -25,7 +26,10 @@ function EarnPage() {
 
   const view = enabled ? <EarnView /> : null
 
-  return <WithConnection>{view}</WithConnection>
+  return <WithConnection>
+    {view}
+    <Survey for="earn" />
+  </WithConnection>
 }
 
 EarnPage.layout = ProductPagesLayout
