+++ pages/multiply.tsx
@@ -1,15 +1,10 @@
-import { Global } from '@emotion/core'
-import { Popover, Sidetab } from '@typeform/embed-react'
-import { useTranslation } from 'next-i18next'
 import { serverSideTranslations } from 'next-i18next/serverSideTranslations'
 import React from 'react'
 
 import { WithConnection } from '../components/connectWallet/ConnectWallet'
 import { ProductPagesLayout } from '../components/Layouts'
 import { MultiplyView } from '../features/multiply/MultiplyView'
-import { useLocalStorage } from '../helpers/useLocalStorage'
-import { useBreakpointIndex } from '../theme/useBreakpointIndex'
-import { useTheme } from '../theme/useThemeUI'
+import { Survey } from 'features/survey'
 
 export const getStaticProps = async ({ locale }: { locale: string }) => ({
   props: {
@@ -17,61 +12,11 @@ export const getStaticProps = async ({ locale }: { locale: string }) => ({
   },
 })
 
-const typeformPopoverButton = '.tf-v1-popover .tf-v1-popover-button'
-
-export function MultiplySurveyButtons() {
-  const breakpoint = useBreakpointIndex()
-  const { theme } = useTheme()
-  const btnColor = theme.colors.link
-  const { t } = useTranslation()
-  const [wasClosed, setWasOpenedAndClosed] = useLocalStorage(
-    'multiply-survey-2022-03-23-was-closed',
-    false,
-  )
-
-  return (
-    <>
-      {breakpoint <= 1 && !wasClosed && (
-        <>
-          {breakpoint === 0 && (
-            <Global
-              styles={() => ({
-                [typeformPopoverButton]: {
-                  bottom: '77px',
-                  right: '14px',
-                },
-              })}
-            />
-          )}
-
-          <Popover
-            id="H52MeocX"
-            buttonColor={btnColor}
-            shareGaInstance={true}
-            onClose={() => setWasOpenedAndClosed(true)}
-          />
-        </>
-      )}
-      {breakpoint > 1 && !wasClosed && (
-        <Sidetab
-          id="H52MeocX"
-          buttonText={t('help-shape-the-future-of-multiply')}
-          buttonColor={btnColor}
-          shareGaInstance={true}
-          onClose={() => setWasOpenedAndClosed(true)}
-        >
-          {t('help-shape-the-future-of-multiply')}
-        </Sidetab>
-      )}
-    </>
-  )
-}
-
 function MultiplyPage() {
   return (
     <WithConnection>
       <MultiplyView />
-      <MultiplySurveyButtons />
+      <Survey for="multiply" />
     </WithConnection>
   )
 }
